# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AIS Catering",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/caterings/{id}/clients/{clientId}/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "caterings orders"
                ],
                "summary": "returns orders of provided client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders for clients",
                        "schema": {
                            "$ref": "#/definitions/swagger.SummaryOrdersResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/clients/{id}/order-status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients orders"
                ],
                "summary": "returns status of order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "order status",
                        "schema": {
                            "$ref": "#/definitions/swagger.OrderStatus"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clients/{id}/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients orders"
                ],
                "summary": "returns orders of provided client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders for clients",
                        "schema": {
                            "$ref": "#/definitions/swagger.SummaryOrdersResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients orders"
                ],
                "summary": "approve user orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Approved orders"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users orders"
                ],
                "summary": "returns orders of provided user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders for user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/swagger.UserOrder"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users orders"
                ],
                "summary": "Returns error or 201 status code if success",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Date query in YYYY-MM-DDT00:00:00Z format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User order",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/swagger.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Order for user",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserOrder"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/orders/{orderId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users orders"
                ],
                "summary": "Returns error or 204 status code if success",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully canceled"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "swagger.ItemsSummaryOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "swagger.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "dishId": {
                    "type": "string"
                }
            }
        },
        "swagger.OrderItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "swagger.OrderRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.Order"
                    }
                }
            }
        },
        "swagger.OrderStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "swagger.SummaryOrder": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.ItemsSummaryOrder"
                    }
                }
            }
        },
        "swagger.SummaryOrdersResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "summary": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.SummaryOrder"
                },
                "summaryTotal": {
                    "type": "integer"
                },
                "userOrders": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.SummaryUserOrder"
                }
            }
        },
        "swagger.SummaryUserOrder": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "floor": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.ItemsSummaryOrder"
                    }
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "swagger.UserOrder": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.OrderItem"
                    }
                },
                "orderId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}